+++ include/trace/events/ext4.h
@@ -322,7 +322,7 @@ TRACE_EVENT(ext4_mark_inode_dirty,
 
 	TP_printk("dev %d,%d ino %lu caller %pS",
 		  MAJOR(__entry->dev), MINOR(__entry->dev),
-		  (unsigned long) __entry->ino, (void *)__entry->ip)
+		  (unsigned long) __entry->ino, (void *)(uintptr_t)__entry->ip)
 );
 
 TRACE_EVENT(ext4_begin_ordered_truncate,
@@ -1778,7 +1778,7 @@ TRACE_EVENT(ext4_journal_start_sb,
 	TP_printk("dev %d,%d blocks %d, rsv_blocks %d, revoke_creds %d,"
 		  " type %d, caller %pS", MAJOR(__entry->dev),
 		  MINOR(__entry->dev), __entry->blocks, __entry->rsv_blocks,
-		  __entry->revoke_creds, __entry->type, (void *)__entry->ip)
+		  __entry->revoke_creds, __entry->type, (void *)(uintptr_t)__entry->ip)
 );
 
 TRACE_EVENT(ext4_journal_start_inode,
@@ -1811,7 +1811,7 @@ TRACE_EVENT(ext4_journal_start_inode,
 		  " type %d, ino %lu, caller %pS", MAJOR(__entry->dev),
 		  MINOR(__entry->dev), __entry->blocks, __entry->rsv_blocks,
 		  __entry->revoke_creds, __entry->type, __entry->ino,
-		  (void *)__entry->ip)
+		  (void *)(uintptr_t)__entry->ip)
 );
 
 TRACE_EVENT(ext4_journal_start_reserved,
@@ -1833,7 +1833,7 @@ TRACE_EVENT(ext4_journal_start_reserved,
 
 	TP_printk("dev %d,%d blocks, %d caller %pS",
 		  MAJOR(__entry->dev), MINOR(__entry->dev),
-		  __entry->blocks, (void *)__entry->ip)
+		  __entry->blocks, (void *)(uintptr_t)__entry->ip)
 );
 
 DECLARE_EVENT_CLASS(ext4__trim,
